python_ai_ocr/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
numpy==1.26.4
opencv-python
paddlepaddle
paddleocr
pytesseract
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
pydantic-settings==2.1.0
supabase==2.5.0